requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
Django==5.2.4
django-ace==1.42.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.4
opencc-python-reimplemented==0.1.7
packaging==25.0
pycparser==2.22
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
scipy==1.15.3
soundfile==0.13.1
sqlparse==0.5.3
srt==3.5.3
sympy==1.14.0
tokenizers==0.19.1
torch==2.1.2
torchaudio==2.1.2
tqdm==4.67.1
transformers==4.40.0
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
vosk==0.3.44
websockets==15.0.1
